:recycle: [Refactor] 이전(7주차) App.tsx 리팩토링 일시 적용 #main
