+++ src/components/calendar/CalendarHeader.tsx
@@ -1,0 +1,27 @@
+import { ChevronLeftIcon, ChevronRightIcon } from '@chakra-ui/icons';
+import { HStack, IconButton, Select } from '@chakra-ui/react';
+
+import { useCalendarView } from '../../hooks/useCalendarView';
+
+export function CalendarHeader() {
+  const { view, setView, navigate } = useCalendarView();
+
+  return (
+    <HStack mx="auto" justifyContent="space-between">
+      <IconButton
+        aria-label="Previous"
+        icon={<ChevronLeftIcon />}
+        onClick={() => navigate('prev')}
+      />
+      <Select
+        aria-label="view"
+        value={view}
+        onChange={(e) => setView(e.target.value as 'week' | 'month')}
+      >
+        <option value="week">Week</option>
+        <option value="month">Month</option>
+      </Select>
+      <IconButton aria-label="Next" icon={<ChevronRightIcon />} onClick={() => navigate('next')} />
+    </HStack>
+  );
+}
