+++ src/components/dialog/Dialog.tsx
@@ -1,0 +1,116 @@
+import {
+  Alert,
+  AlertDialog,
+  AlertDialogBody,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogOverlay,
+  AlertIcon,
+  AlertTitle,
+  Box,
+  Button,
+  CloseButton,
+  Text,
+  VStack,
+} from '@chakra-ui/react';
+import { useRef } from 'react';
+
+import { useDialog } from '../../hooks/useDialog';
+import { useEventForm } from '../../hooks/useEventForm';
+import { useEventOperations } from '../../hooks/useEventOperations';
+import { useNotifications } from '../../hooks/useNotifications';
+
+export function Dialog() {
+  const {
+    title,
+    date,
+    startTime,
+    endTime,
+    description,
+    location,
+    category,
+    isRepeating,
+    repeatType,
+    repeatInterval,
+    repeatEndDate,
+    notificationTime,
+    editingEvent,
+  } = useEventForm();
+  const { events } = useEventOperations(false);
+  const { saveEvent } = useEventOperations(Boolean(editingEvent), () => null);
+  const { notifications, setNotifications } = useNotifications(events);
+  const { isOverlapDialogOpen, setIsOverlapDialogOpen, overlappingEvents } = useDialog();
+  const cancelRef = useRef<HTMLButtonElement>(null);
+  return (
+    <>
+      <AlertDialog
+        isOpen={isOverlapDialogOpen}
+        leastDestructiveRef={cancelRef}
+        onClose={() => setIsOverlapDialogOpen(false)}
+      >
+        <AlertDialogOverlay>
+          <AlertDialogContent>
+            <AlertDialogHeader fontSize="lg" fontWeight="bold">
+              일정 겹침 경고
+            </AlertDialogHeader>
+            <AlertDialogBody>
+              다음 일정과 겹칩니다:
+              {overlappingEvents.map((event) => (
+                <Text key={event.id}>
+                  {event.title} ({event.date} {event.startTime}-{event.endTime})
+                </Text>
+              ))}
+              계속 진행하시겠습니까?
+            </AlertDialogBody>
+            <AlertDialogFooter>
+              <Button ref={cancelRef} onClick={() => setIsOverlapDialogOpen(false)}>
+                취소
+              </Button>
+              <Button
+                colorScheme="red"
+                onClick={() => {
+                  setIsOverlapDialogOpen(false);
+                  saveEvent({
+                    id: editingEvent ? editingEvent.id : undefined,
+                    title,
+                    date,
+                    startTime,
+                    endTime,
+                    description,
+                    location,
+                    category,
+                    repeat: {
+                      type: isRepeating ? repeatType : 'none',
+                      interval: repeatInterval,
+                      endDate: repeatEndDate || undefined,
+                    },
+                    notificationTime,
+                  });
+                }}
+                ml={3}
+              >
+                계속 진행
+              </Button>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialogOverlay>
+      </AlertDialog>
+      {notifications.length > 0 && (
+        <VStack position="fixed" top={4} right={4} spacing={2} align="flex-end">
+          {notifications.map((notification, index) => (
+            <Alert key={index} status="info" variant="solid" width="auto">
+              <AlertIcon />
+              <Box flex="1">
+                <AlertTitle fontSize="sm">{notification.message}</AlertTitle>
+              </Box>
+              <CloseButton
+                onClick={() => setNotifications((prev) => prev.filter((_, i) => i !== index))}
+              />
+            </Alert>
+          ))}
+        </VStack>
+      )}
+    </>
+  );
+}
