+++ src/hooks/useEventForm.ts
@@ -1,3 +1,4 @@
+import { atom, useAtom } from 'jotai';
 import { ChangeEvent, useState } from 'react';
 
 import { Event, RepeatType } from '../types';
@@ -5,21 +6,34 @@ import { getTimeErrorMessage } from '../utils/timeValidation';
 
 type TimeErrorRecord = Record<'startTimeError' | 'endTimeError', string | null>;
 
-export const useEventForm = (initialEvent?: Event) => {
-  const [title, setTitle] = useState(initialEvent?.title || '');
-  const [date, setDate] = useState(initialEvent?.date || '');
-  const [startTime, setStartTime] = useState(initialEvent?.startTime || '');
-  const [endTime, setEndTime] = useState(initialEvent?.endTime || '');
-  const [description, setDescription] = useState(initialEvent?.description || '');
-  const [location, setLocation] = useState(initialEvent?.location || '');
-  const [category, setCategory] = useState(initialEvent?.category || '');
-  const [isRepeating, setIsRepeating] = useState(initialEvent?.repeat.type !== 'none');
-  const [repeatType, setRepeatType] = useState<RepeatType>(initialEvent?.repeat.type || 'none');
-  const [repeatInterval, setRepeatInterval] = useState(initialEvent?.repeat.interval || 1);
-  const [repeatEndDate, setRepeatEndDate] = useState(initialEvent?.repeat.endDate || '');
-  const [notificationTime, setNotificationTime] = useState(initialEvent?.notificationTime || 10);
+const titleAtom = atom('');
+const dateAtom = atom('');
+const startTimeAtom = atom('');
+const endTimeAtom = atom('');
+const descriptionAtom = atom('');
+const locationAtom = atom('');
+const categoryAtom = atom('');
+const isRepeatingAtom = atom(false);
+const repeatTypeAtom = atom<RepeatType>('none');
+const repeatIntervalAtom = atom(1);
+const repeatEndDateAtom = atom('');
+const notificationTimeAtom = atom(10);
+const editingEventAtom = atom<Event | null>(null);
 
-  const [editingEvent, setEditingEvent] = useState<Event | null>(null);
+export const useEventForm = () => {
+  const [title, setTitle] = useAtom(titleAtom);
+  const [date, setDate] = useAtom(dateAtom);
+  const [startTime, setStartTime] = useAtom(startTimeAtom);
+  const [endTime, setEndTime] = useAtom(endTimeAtom);
+  const [description, setDescription] = useAtom(descriptionAtom);
+  const [location, setLocation] = useAtom(locationAtom);
+  const [category, setCategory] = useAtom(categoryAtom);
+  const [isRepeating, setIsRepeating] = useAtom(isRepeatingAtom);
+  const [repeatType, setRepeatType] = useAtom(repeatTypeAtom);
+  const [repeatInterval, setRepeatInterval] = useAtom(repeatIntervalAtom);
+  const [repeatEndDate, setRepeatEndDate] = useAtom(repeatEndDateAtom);
+  const [notificationTime, setNotificationTime] = useAtom(notificationTimeAtom);
+  const [editingEvent, setEditingEvent] = useAtom(editingEventAtom);
 
   const [{ startTimeError, endTimeError }, setTimeError] = useState<TimeErrorRecord>({
     startTimeError: null,
