+++ src/hooks/useSearch.ts
@@ -1,10 +1,13 @@
-import { useMemo, useState } from 'react';
+import { atom, useAtom } from 'jotai';
+import { useMemo } from 'react';
 
 import { Event } from '../types';
 import { getFilteredEvents } from '../utils/eventUtils';
 
+const searchTermAtom = atom('');
+
 export const useSearch = (events: Event[], currentDate: Date, view: 'week' | 'month') => {
-  const [searchTerm, setSearchTerm] = useState('');
+  const [searchTerm, setSearchTerm] = useAtom(searchTermAtom);
 
   const filteredEvents = useMemo(() => {
     return getFilteredEvents(events, searchTerm, currentDate, view);
